requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asyncio==3.4.3
attrs==22.2.0
black==24.4.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
flake8==7.0.0
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huepy==1.2.1
identify==2.5.36
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
instagrapi==2.1.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
maxminddb==2.6.1
mccabe==0.7.0
mdurl==0.1.2
mock==5.1.0
moviepy==1.0.3
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.9.0
numpy==1.26.4
orjson==3.10.3
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
proglog==0.1.10
pyasn1==0.6.0
pycares==4.4.0
pycodestyle==2.11.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
Pygments==2.18.0
pymongo==4.7.2
PySocks==1.7.1
pytest==8.2.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
requests-toolbelt==1.0.0
responses==0.25.0
rich==13.7.1
rsa==4.9
schedule==1.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
